finished the refactoring
